requirements.txt
altgraph==0.17
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Django==3.1.6
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-request==1.5.6
djangorestframework==3.12.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.3
numpy==1.20.0
oauthlib==3.1.0
pandas==1.2.1
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pyperclip==1.8.1
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pywin32-ctypes==0.2.0
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
twilio==6.51.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
